Refactor for wide dates csv (#601)
### This PR

Doesn't introduce behavior changes, just refactoring.

* Changes `update_data_public_head` to always get data as arguments, making dependency explicit. Also I want `cli.data.update` to have the data so it can write the wide date csv in my next PR.
* Factor out `fips_index_geo_data` so it can be reused
* in `timeseries_test.py` give every row a unique FIPS value
* move `generate_field_summary` so it can be imported by `TimeseriesDataset` without creating an import loop.
